backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.32
alembic==1.13.2
pymysql==1.1.1
python-dotenv==1.0.1
httpx==0.27.2
python-jose[cryptography]==3.3.0
cryptography==43.0.1
passlib[bcrypt]==1.7.4
pydantic==2.8.2
python-multipart==0.0.9
structlog==24.4.0